added request handler to application

+++ back-end/src/api/controllers/application_controller.py
@@ -12,43 +12,68 @@ def get_applications():
     """
         This is the endpoint returning application list
     """
-    api_response = application_service.get_applications()
-    return api_response
+    req_handler_response = request_handler.validate_token(request)
+
+    if req_handler_response['success']:
+        api_response = application_service.get_applications()
+        return api_response
+
+    return req_handler_response
 
 @app.route(api_version + '/<application_id>', methods=['GET'])
 def get_application(application_id):
     """
         This is the endpoint returning a single application with the given id
     """
-    api_response = application_service.get_application(application_id)
-    return api_response
+    req_handler_response = request_handler.validate_token(request)
+
+    if req_handler_response['success']:
+        api_response = application_service.get_application(application_id)
+        return api_response
+
+    return req_handler_response    
 
 @app.route(api_version + '/', methods=['POST'])
 def insert_application():
     """
         This is the endpoint for creating a new application
     """
-    data = request.get_json()
-    api_response = application_service.create_application(data)
-    return api_response
+    req_handler_response = request_handler.validate_token(request)
+
+    if req_handler_response['success']:
+        data = request.get_json()
+        api_response = application_service.insert_application(data)
+        return api_response
+
+    return req_handler_response
 
 @app.route(api_version + '/', methods=['DELETE'])
 def delete_application():
     """
         This is endpoint for deleting an application 
     """
-    data = request.get_json()
-    api_response = application_service.delete_application(data)
-    return api_response
+    req_handler_response = request_handler.validate_token(request)
+
+    if req_handler_response['success']:
+        data = request.get_json()
+        api_response = application_service.delete_application(data)
+        return api_response
+
+    return req_handler_response
 
 @app.route(api_version + '/', methods=['PUT'])
 def update_application():
     """
         This is endpoint for updating an application 
     """
-    data = request.get_json()
-    api_response = application_service.update_application(data)
-    return api_response
+    req_handler_response = request_handler.validate_token(request)
+
+    if req_handler_response['success']:
+        data = request.get_json()
+        api_response = application_service.update_application(data)
+        return api_response
+
+    return req_handler_response
 
 
 @app.route(api_version + '/status', methods=['PUT'])
